Prompts/requirements.txt
# LangChain Core
langchain
langchain-core

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# Streamlit for Web UI
streamlit


